requirements.txt
pydicom>=2.4.0
torch>=2.1.0
torchvision>=0.16.0
fastapi>=0.109.0
uvicorn>=0.27.0
numpy>=1.26.0
Pillow>=10.1.0
python-multipart>=0.0.6
pytest>=7.4.0
python-dotenv>=1.0.0